adapted registration form to new UI pattern
