+++ src/app/admin/products/modals/InsightsModal.jsx
